backend/requirements.txt
Django
djangorestframework
gunicorn
zmq
